A late-bound helper has priority over variable
This has already been the case for late-bound helpers _with arguments_
since they were easily recognised as helper calls. However, helper
resolution was not attempted for simple paths. E.g. for `{{foo}}`, if
"foo" was late-bound, we never attempted to invoke it, we only looked
for "foo" in the data.